+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## [Unreleased]
 
+
 ## [1.1.0]
 
 - Added support for custom html for icons in command items.
+- Added support for 'visible' property in command items to conditionally display items.
 
 
 ## [1.0.0]